requirements.txt
--extra-index-url https://asia-northeast1-python.pkg.dev/meet-notify/google-apps-meet/simple/
blinker==1.7.0
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
cloudevents==1.10.1
deprecation==2.1.0
docutils==0.20.1
Flask==3.0.0
functions-framework==3.5.0
google-api-core==2.14.0
google-apps-meet==0.0.0
google-auth==2.25.0
google-auth-oauthlib==1.1.0
google-cloud-pubsub==2.18.4
googleapis-common-protos==1.61.0
grpc-google-iam-v1==0.12.7
grpcio==1.59.3
grpcio-status==1.59.3
gunicorn==21.2.0
idna==3.6
importlib-metadata==7.0.0
itsdangerous==2.1.2
jaraco.classes==3.3.0
Jinja2==3.1.2
keyring==24.3.0
keyrings.google-artifactregistry-auth==1.1.2
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
more-itertools==10.1.0
nh3==0.2.14
oauthlib==3.2.2
packaging==23.2
pip==23.3.1
pkginfo==1.9.6
pluggy==1.3.0
proto-plus==1.22.3
protobuf==4.25.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
Pygments==2.17.2
python-dotenv==1.0.0
readme-renderer==42.0
requests==2.31.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==13.7.0
rsa==4.9
setuptools==68.2.2
twine==4.0.2
urllib3==2.1.0
watchdog==3.0.0
Werkzeug==3.0.1
wheel==0.42.0
zipp==3.17.0
